[tmp] restrict tests
test_annex_ssh stalled at

    https://travis-ci.org/github/datalad/datalad/builds/706678114

760c8075f (TST: annexrepo: Clear repo instances when testing shared
connections, 2020-05-12) removed the skip, hoping it had resolved it.
It pointed to two travis runs, where the one right before that commit
stalled, but perhaps it's just flaky and that conclusion is wrong.  Or
perhaps it's only when a Docker target is used.

I can't get it to hang in a Xenial VM though (on this branch, with
latest annex commit).

So, try again with Docker target to at least get sense of whether it's
an intermittent stall.